refactor: move attribute type to enums file
